+ Added GROWING

+++ src/gameproject/Circle.java
@@ -24,24 +24,24 @@ public class Circle
     /**
      * The radius of the cell.
      */
-    protected int radius;
+    protected double radius;
     
     public int getCenterX()
     {
-        return x + radius;
+        return x + (int)radius;
     }
     
     public int getCenterY()
     {
-        return y + radius;
+        return y + (int)radius;
     }
     
     public boolean collidesWith(Circle other)
     {
-        float dx = getCenterX() - other.getCenterX();
-        float dy = getCenterY() - other.getCenterY();
-        float distance = dx * dx + dy * dy;
-        float radiusSum = radius + other.radius;
+        double dx = getCenterX() - other.getCenterX();
+        double dy = getCenterY() - other.getCenterY();
+        double distance = dx * dx + dy * dy;
+        double radiusSum = radius + other.radius;
         return distance < radiusSum * radiusSum;
     }
 }
